% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{get_planet}
\alias{entries}
\alias{get_film}
\alias{get_person}
\alias{get_planet}
\alias{get_species}
\alias{get_starship}
\alias{get_vehicle}
\title{Retrieve Data about a Specific DB Entry}
\usage{
get_planet(id, parse_result = FALSE, ...)

get_species(id, parse_result = FALSE, ...)

get_film(id, parse_result = FALSE, ...)

get_vehicle(id, parse_result = FALSE, ...)

get_starship(id, parse_result = FALSE, ...)

get_person(id, parse_result = FALSE, ...)
}
\arguments{
\item{id}{the numeric ID of the (planet, species, etc)
you want to retrieve. Note that IDs are not necessarily sequential;
IDs 3, 4 and 6 may exist, while ID 5 may error.}

\item{parse_result}{whether to parse and simplify the data before
returning it. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET function.}
}
\description{
These functions return a specific DB entry's information - where
that entry can be a planet, species, vehicle, so on and so forth.
}
\examples{
#Retrieve an object by ID
looking_for_data_in_alderaan_places <- get_planet(2)

}
\seealso{
retrieving entire classes of entries with \code{\link{all_entries}}.
}

