% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schemas.R
\name{get_planet_schema}
\alias{get_film_schema}
\alias{get_person_schema}
\alias{get_planet_schema}
\alias{get_species_schema}
\alias{get_starship_schema}
\alias{get_vehicle_schema}
\alias{schemas}
\title{Retrieve the schema specifications for different object classes}
\usage{
get_planet_schema(parse_result = FALSE, ...)

get_species_schema(parse_result = FALSE, ...)

get_film_schema(parse_result = FALSE, ...)

get_vehicle_schema(parse_result = FALSE, ...)

get_starship_schema(parse_result = FALSE, ...)

get_person_schema(parse_result = FALSE, ...)
}
\arguments{
\item{parse_result}{whether to simplify the results before they
are returned. This is FALSE by default.}

\item{...}{additional arguments to pass to httr's GET function,
such as timeout settings or user agents.}
}
\value{
a list-based representation of the relevant JSON schema.
}
\description{
*_schema functions grab the schema specifications for each
class of objects in the SWAPI database. This can be used to predict what
structure data returned from the API will take.
}
\examples{
get_planet_schema()
}
\seealso{
\code{\link{entries}} for retrieving data about
individual planets, schemas, etc, or \code{\link{all_entries}}
for retrieving all of the data in a particular category.
}

