% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all.R, R/entries.R
\name{get_all_species}
\alias{entries}
\alias{get_all_films}
\alias{get_all_people}
\alias{get_all_species}
\alias{get_all_starships}
\alias{get_all_vehicles}
\alias{get_film}
\alias{get_person}
\alias{get_planet}
\alias{get_species}
\alias{get_starship}
\alias{get_vehicle}
\title{Retrieve the data associated with a specific entry
(planet, species, etc)}
\usage{
get_all_species(parse_result = FALSE, ...)

get_all_films(parse_result = FALSE, ...)

get_all_vehicles(parse_result = FALSE, ...)

get_all_starships(parse_result = FALSE, ...)

get_all_people(parse_result = FALSE, ...)

get_planet(id, parse_result = FALSE, ...)

get_species(id, parse_result = FALSE, ...)

get_film(id, parse_result = FALSE, ...)

get_vehicle(id, parse_result = FALSE, ...)

get_starship(id, parse_result = FALSE, ...)

get_person(id, parse_result = FALSE, ...)
}
\description{
Retrieve the data associated with a specific entry
(planet, species, etc)
}

