% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entries.R
\name{get_planet}
\alias{entries}
\alias{get_film}
\alias{get_planet}
\alias{get_species}
\alias{get_starship}
\alias{get_vehicle}
\title{Retrieve the data associated with a specific object
(planet, species, etc)}
\usage{
get_planet(id, parse_result = FALSE, ...)

get_species(id, parse_result = FALSE, ...)

get_film(id, parse_result = FALSE, ...)

get_vehicle(id, parse_result = FALSE, ...)

get_starship(id, parse_result = FALSE, ...)
}
\description{
Retrieve the data associated with a specific object
(planet, species, etc)
}

